service-a/requirements/production.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
flask-restful==0.3.9
flask-migrate==3.1.0
flask-seeder==1.2.0
faker==8.11.0
psycopg2-binary==2.9.1
requests==2.26.0
python-dotenv==0.19.0
cerberus==1.3.4
pyjwt==2.1.0
validate-cpf==0.1.2
